Move all flag related features to seperate file
